requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.12.0
python-dotenv==1.0.0
elevenlabs>=2.32.0
requests==2.31.0
git+https://github.com/RVC-Project/Retrieval-based-Voice-Conversion
scipy>=1.11.0
python-multipart==0.0.6
aiofiles==23.2.1
